refactor: move rest of proofs to `penumbra-shielded-pool`
