+++ crates/core/component/shielded-pool/src/output.rs
@@ -1,7 +1,9 @@
 mod action;
 mod plan;
+mod proof;
 mod view;
 
 pub use action::{Body, Output};
 pub use plan::OutputPlan;
+pub use proof::{OutputCircuit, OutputProof};
 pub use view::OutputView;
