+++ crates/core/component/shielded-pool/src/spend.rs
@@ -1,7 +1,9 @@
 mod action;
 mod plan;
+mod proof;
 mod view;
 
 pub use action::{Body, Spend};
 pub use plan::SpendPlan;
+pub use proof::{SpendCircuit, SpendProof};
 pub use view::SpendView;
