+++ crates/core/crypto/src/proofs/groth16.rs
@@ -1,12 +1,5 @@
-mod gadgets;
-mod nullifier_derivation;
-mod output;
-mod spend;
 mod traits;
 
-pub use nullifier_derivation::{NullifierDerivationCircuit, NullifierDerivationProof};
-pub use output::{OutputCircuit, OutputProof};
-pub use spend::{SpendCircuit, SpendProof};
 pub use traits::{ParameterSetup, ProvingKeyExt, VerifyingKeyExt};
 
 /// The length of our Groth16 proofs in bytes.
@@ -51,608 +44,6 @@ mod tests {
         any::<[u8; 32]>()
             .prop_map(|bytes| Fr::from_le_bytes_mod_order(&bytes[..]))
             .boxed()
-    }
-
-    proptest! {
-    #![proptest_config(ProptestConfig::with_cases(2))]
-    #[test]
-    fn nullifier_derivation_proof_happy_path(seed_phrase_randomness in any::<[u8; 32]>(), value_amount in 2..200u64) {
-            let (pk, vk) = NullifierDerivationCircuit::generate_prepared_test_parameters();
-
-            let mut rng = OsRng;
-
-            let seed_phrase = SeedPhrase::from_randomness(seed_phrase_randomness);
-            let sk_sender = SpendKey::from_seed_phrase(seed_phrase, 0);
-            let fvk_sender = sk_sender.full_viewing_key();
-            let ivk_sender = fvk_sender.incoming();
-            let (sender, _dtk_d) = ivk_sender.payment_address(0u32.into());
-
-            let value_to_send = Value {
-                amount: value_amount.into(),
-                asset_id: asset::Cache::with_known_assets().get_unit("upenumbra").unwrap().id(),
-            };
-
-            let note = Note::generate(&mut rng, &sender, value_to_send);
-            let note_commitment = note.commit();
-            let nk = *sk_sender.nullifier_key();
-            let mut sct = tct::Tree::new();
-
-            sct.insert(tct::Witness::Keep, note_commitment).unwrap();
-            let state_commitment_proof = sct.witness(note_commitment).unwrap();
-            let position = state_commitment_proof.position();
-            let nullifier = Nullifier::derive(&nk, state_commitment_proof.position(), &note_commitment);
-
-                let proof = NullifierDerivationProof::prove(
-                    &mut rng,
-                    &pk,
-                    position,
-                    note.clone(),
-                    nk,
-                    nullifier,
-                )
-                .expect("can create proof");
-
-                let proof_result = proof.verify(&vk, position, note, nullifier);
-
-                assert!(proof_result.is_ok());
-        }
-    }
-
-    proptest! {
-    #![proptest_config(ProptestConfig::with_cases(2))]
-    #[test]
-    fn output_proof_happy_path(seed_phrase_randomness in any::<[u8; 32]>(), v_blinding in fr_strategy(), value_amount in 2..200u64) {
-            let (pk, vk) = OutputCircuit::generate_prepared_test_parameters();
-
-            let mut rng = OsRng;
-
-            let seed_phrase = SeedPhrase::from_randomness(seed_phrase_randomness);
-            let sk_recipient = SpendKey::from_seed_phrase(seed_phrase, 0);
-            let fvk_recipient = sk_recipient.full_viewing_key();
-            let ivk_recipient = fvk_recipient.incoming();
-            let (dest, _dtk_d) = ivk_recipient.payment_address(0u32.into());
-
-            let value_to_send = Value {
-                amount: value_amount.into(),
-                asset_id: asset::Cache::with_known_assets().get_unit("upenumbra").unwrap().id(),
-            };
-
-            let note = Note::generate(&mut rng, &dest, value_to_send);
-            let note_commitment = note.commit();
-            let balance_commitment = (-Balance::from(value_to_send)).commit(v_blinding);
-
-            let blinding_r = Fq::rand(&mut OsRng);
-            let blinding_s = Fq::rand(&mut OsRng);
-            let proof = OutputProof::prove(
-                blinding_r,
-                blinding_s,
-                &pk,
-                note,
-                v_blinding,
-                balance_commitment,
-                note_commitment,
-            )
-            .expect("can create proof");
-            let serialized_proof: pb::ZkOutputProof = proof.into();
-
-            let deserialized_proof = OutputProof::try_from(serialized_proof).expect("can deserialize proof");
-            let proof_result = deserialized_proof.verify(&vk, balance_commitment, note_commitment);
-
-            assert!(proof_result.is_ok());
-        }
-    }
-
-    proptest! {
-    #![proptest_config(ProptestConfig::with_cases(2))]
-    #[test]
-    fn output_proof_verification_note_commitment_integrity_failure(seed_phrase_randomness in any::<[u8; 32]>(), v_blinding in fr_strategy(), value_amount in 2..200u64, note_blinding in fq_strategy()) {
-        let (pk, vk) = OutputCircuit::generate_prepared_test_parameters();
-        let mut rng = OsRng;
-
-        let seed_phrase = SeedPhrase::from_randomness(seed_phrase_randomness);
-        let sk_recipient = SpendKey::from_seed_phrase(seed_phrase, 0);
-        let fvk_recipient = sk_recipient.full_viewing_key();
-        let ivk_recipient = fvk_recipient.incoming();
-        let (dest, _dtk_d) = ivk_recipient.payment_address(0u32.into());
-
-        let value_to_send = Value {
-            amount: value_amount.into(),
-            asset_id: asset::Cache::with_known_assets().get_unit("upenumbra").unwrap().id(),
-        };
-
-        let note = Note::generate(&mut rng, &dest, value_to_send);
-        let note_commitment = note.commit();
-        let balance_commitment = (-Balance::from(value_to_send)).commit(v_blinding);
-
-        let blinding_r = Fq::rand(&mut OsRng);
-        let blinding_s = Fq::rand(&mut OsRng);
-        let proof = OutputProof::prove(
-            blinding_r,
-            blinding_s,
-            &pk,
-            note.clone(),
-            v_blinding,
-            balance_commitment,
-            note_commitment,
-        )
-        .expect("can create proof");
-
-        let incorrect_note_commitment = note::commitment(
-            note_blinding,
-            value_to_send,
-            note.diversified_generator(),
-            note.transmission_key_s(),
-            note.clue_key(),
-        );
-
-        let proof_result = proof.verify(&vk, balance_commitment, incorrect_note_commitment);
-
-        assert!(proof_result.is_err());
-    }
-    }
-
-    proptest! {
-        #![proptest_config(ProptestConfig::with_cases(2))]
-        #[test]
-    fn output_proof_verification_balance_commitment_integrity_failure(seed_phrase_randomness in any::<[u8; 32]>(), v_blinding in fr_strategy(), value_amount in 2..200u64, incorrect_v_blinding in fr_strategy()) {
-        let (pk, vk) = OutputCircuit::generate_prepared_test_parameters();
-        let mut rng = OsRng;
-
-        let seed_phrase = SeedPhrase::from_randomness(seed_phrase_randomness);
-        let sk_recipient = SpendKey::from_seed_phrase(seed_phrase, 0);
-        let fvk_recipient = sk_recipient.full_viewing_key();
-        let ivk_recipient = fvk_recipient.incoming();
-        let (dest, _dtk_d) = ivk_recipient.payment_address(0u32.into());
-
-        let value_to_send = Value {
-            amount: value_amount.into(),
-            asset_id: asset::Cache::with_known_assets().get_unit("upenumbra").unwrap().id(),
-        };
-
-        let note = Note::generate(&mut rng, &dest, value_to_send);
-        let note_commitment = note.commit();
-        let balance_commitment = (-Balance::from(value_to_send)).commit(v_blinding);
-
-        let blinding_r = Fq::rand(&mut OsRng);
-        let blinding_s = Fq::rand(&mut OsRng);
-        let proof = OutputProof::prove(
-            blinding_r,
-            blinding_s,
-            &pk,
-            note,
-            v_blinding,
-            balance_commitment,
-            note_commitment,
-        )
-        .expect("can create proof");
-
-        let incorrect_balance_commitment = (-Balance::from(value_to_send)).commit(incorrect_v_blinding);
-
-        let proof_result = proof.verify(&vk, incorrect_balance_commitment, note_commitment);
-
-        assert!(proof_result.is_err());
-    }
-        }
-
-    proptest! {
-        #![proptest_config(ProptestConfig::with_cases(2))]
-    #[test]
-    /// Check that the `SpendProof` verification succeeds.
-    fn spend_proof_verification_success(seed_phrase_randomness in any::<[u8; 32]>(), spend_auth_randomizer in fr_strategy(), value_amount in 2..2000000000u64, num_commitments in 1..2000u64, v_blinding in fr_strategy()) {
-        let (pk, vk) = SpendCircuit::generate_prepared_test_parameters();
-        let mut rng = OsRng;
-
-        let seed_phrase = SeedPhrase::from_randomness(seed_phrase_randomness);
-        let sk_sender = SpendKey::from_seed_phrase(seed_phrase, 0);
-        let fvk_sender = sk_sender.full_viewing_key();
-        let ivk_sender = fvk_sender.incoming();
-        let (sender, _dtk_d) = ivk_sender.payment_address(0u32.into());
-
-        let value_to_send = Value {
-            amount: value_amount.into(),
-            asset_id: asset::Cache::with_known_assets().get_unit("upenumbra").unwrap().id(),
-        };
-
-        let note = Note::generate(&mut rng, &sender, value_to_send);
-        let note_commitment = note.commit();
-        let rsk = sk_sender.spend_auth_key().randomize(&spend_auth_randomizer);
-        let nk = *sk_sender.nullifier_key();
-        let ak: VerificationKey<SpendAuth> = sk_sender.spend_auth_key().into();
-        let mut sct = tct::Tree::new();
-
-        // Next, we simulate the case where the SCT is not empty by adding `num_commitments`
-        // unrelated items in the SCT.
-        for _ in 0..num_commitments {
-            let random_note_commitment = Note::generate(&mut rng, &sender, value_to_send).commit();
-            sct.insert(tct::Witness::Keep, random_note_commitment).unwrap();
-        }
-
-        sct.insert(tct::Witness::Keep, note_commitment).unwrap();
-        let anchor = sct.root();
-        let state_commitment_proof = sct.witness(note_commitment).unwrap();
-        let balance_commitment = value_to_send.commit(v_blinding);
-        let rk: VerificationKey<SpendAuth> = rsk.into();
-        let nf = Nullifier::derive(&nk, state_commitment_proof.position(), &note_commitment);
-
-        let blinding_r = Fq::rand(&mut OsRng);
-        let blinding_s = Fq::rand(&mut OsRng);
-        let proof = SpendProof::prove(
-            blinding_r,
-            blinding_s,
-            &pk,
-            state_commitment_proof,
-            note,
-            v_blinding,
-            spend_auth_randomizer,
-            ak,
-            nk,
-            anchor,
-            balance_commitment,
-            nf,
-            rk,
-        )
-        .expect("can create proof");
-
-        let proof_result = proof.verify(&vk, anchor, balance_commitment, nf, rk);
-        assert!(proof_result.is_ok());
-    }
-    }
-
-    proptest! {
-        #![proptest_config(ProptestConfig::with_cases(2))]
-    #[test]
-    /// Check that the `SpendProof` verification fails when using an incorrect
-    /// TCT root (`anchor`).
-    fn spend_proof_verification_merkle_path_integrity_failure(seed_phrase_randomness in any::<[u8; 32]>(), spend_auth_randomizer in fr_strategy(), value_amount in 2..200u64, v_blinding in fr_strategy()) {
-        let (pk, vk) = SpendCircuit::generate_prepared_test_parameters();
-        let mut rng = OsRng;
-
-        let seed_phrase = SeedPhrase::from_randomness(seed_phrase_randomness);
-        let sk_sender = SpendKey::from_seed_phrase(seed_phrase, 0);
-        let fvk_sender = sk_sender.full_viewing_key();
-        let ivk_sender = fvk_sender.incoming();
-        let (sender, _dtk_d) = ivk_sender.payment_address(0u32.into());
-
-        let value_to_send = Value {
-            amount: value_amount.into(),
-            asset_id: asset::Cache::with_known_assets().get_unit("upenumbra").unwrap().id(),
-        };
-
-        let note = Note::generate(&mut rng, &sender, value_to_send);
-        let note_commitment = note.commit();
-        let rsk = sk_sender.spend_auth_key().randomize(&spend_auth_randomizer);
-        let nk = *sk_sender.nullifier_key();
-        let ak: VerificationKey<SpendAuth> = sk_sender.spend_auth_key().into();
-        let mut sct = tct::Tree::new();
-        let incorrect_anchor = sct.root();
-        sct.insert(tct::Witness::Keep, note_commitment).unwrap();
-        let anchor = sct.root();
-        let state_commitment_proof = sct.witness(note_commitment).unwrap();
-        let balance_commitment = value_to_send.commit(v_blinding);
-        let rk: VerificationKey<SpendAuth> = rsk.into();
-        let nf = Nullifier::derive(&nk, 0.into(), &note_commitment);
-
-        let blinding_r = Fq::rand(&mut OsRng);
-        let blinding_s = Fq::rand(&mut OsRng);
-        let proof = SpendProof::prove(
-            blinding_r,
-            blinding_s,
-            &pk,
-            state_commitment_proof,
-            note,
-            v_blinding,
-            spend_auth_randomizer,
-            ak,
-            nk,
-            anchor,
-            balance_commitment,
-            nf,
-            rk,
-        )
-        .expect("can create proof");
-
-        let proof_result = proof.verify(&vk, incorrect_anchor, balance_commitment, nf, rk);
-        assert!(proof_result.is_err());
-    }
-    }
-
-    proptest! {
-            #![proptest_config(ProptestConfig::with_cases(2))]
-            #[should_panic]
-        #[test]
-        /// Check that the `SpendProof` verification fails when the diversified address is wrong.
-        fn spend_proof_verification_diversified_address_integrity_failure(seed_phrase_randomness in any::<[u8; 32]>(), incorrect_seed_phrase_randomness in any::<[u8; 32]>(), spend_auth_randomizer in fr_strategy(), value_amount in 2..200u64, v_blinding in fr_strategy()) {
-            let (pk, vk) = SpendCircuit::generate_prepared_test_parameters();
-            let mut rng = OsRng;
-
-            let seed_phrase = SeedPhrase::from_randomness(seed_phrase_randomness);
-            let sk_sender = SpendKey::from_seed_phrase(seed_phrase, 0);
-
-            let wrong_seed_phrase = SeedPhrase::from_randomness(incorrect_seed_phrase_randomness);
-            let wrong_sk_sender = SpendKey::from_seed_phrase(wrong_seed_phrase, 0);
-            let wrong_fvk_sender = wrong_sk_sender.full_viewing_key();
-            let wrong_ivk_sender = wrong_fvk_sender.incoming();
-            let (wrong_sender, _dtk_d) = wrong_ivk_sender.payment_address(1u32.into());
-
-            let value_to_send = Value {
-                amount: value_amount.into(),
-                asset_id: asset::Cache::with_known_assets().get_unit("upenumbra").unwrap().id(),
-            };
-
-            let note = Note::generate(&mut rng, &wrong_sender, value_to_send);
-
-            let note_commitment = note.commit();
-            let rsk = sk_sender.spend_auth_key().randomize(&spend_auth_randomizer);
-            let nk = *sk_sender.nullifier_key();
-            let ak = sk_sender.spend_auth_key().into();
-            let mut sct = tct::Tree::new();
-            sct.insert(tct::Witness::Keep, note_commitment).unwrap();
-            let anchor = sct.root();
-            let state_commitment_proof = sct.witness(note_commitment).unwrap();
-            let balance_commitment = value_to_send.commit(v_blinding);
-            let rk: VerificationKey<SpendAuth> = rsk.into();
-            let nf = Nullifier::derive(&nk, 0.into(), &note_commitment);
-
-            // Note that this will blow up in debug mode as the constraint
-            // system is unsatisified (ark-groth16 has a debug check for this).
-            // In release mode the proof will be created, but will fail to verify.
-            let blinding_r = Fq::rand(&mut OsRng);
-            let blinding_s = Fq::rand(&mut OsRng);
-            let proof = SpendProof::prove(
-                blinding_r,
-                blinding_s,
-                &pk,
-                state_commitment_proof,
-                note,
-                v_blinding,
-                spend_auth_randomizer,
-                ak,
-                nk,
-                anchor,
-                balance_commitment,
-                nf,
-                rk,
-            ).expect("can create proof in release mode");
-
-            proof.verify(&vk, anchor, balance_commitment, nf, rk).expect("boom");
-        }
-    }
-
-    proptest! {
-            #![proptest_config(ProptestConfig::with_cases(2))]
-        #[test]
-        /// Check that the `SpendProof` verification fails, when using an
-        /// incorrect nullifier.
-        fn spend_proof_verification_nullifier_integrity_failure(seed_phrase_randomness in any::<[u8; 32]>(), spend_auth_randomizer in fr_strategy(), value_amount in 2..200u64, v_blinding in fr_strategy()) {
-            let (pk, vk) = SpendCircuit::generate_prepared_test_parameters();
-            let mut rng = OsRng;
-
-            let seed_phrase = SeedPhrase::from_randomness(seed_phrase_randomness);
-            let sk_sender = SpendKey::from_seed_phrase(seed_phrase, 0);
-            let fvk_sender = sk_sender.full_viewing_key();
-            let ivk_sender = fvk_sender.incoming();
-            let (sender, _dtk_d) = ivk_sender.payment_address(0u32.into());
-
-            let value_to_send = Value {
-                amount: value_amount.into(),
-                asset_id: asset::Cache::with_known_assets().get_unit("upenumbra").unwrap().id(),
-            };
-
-            let note = Note::generate(&mut rng, &sender, value_to_send);
-            let note_commitment = note.commit();
-            let rsk = sk_sender.spend_auth_key().randomize(&spend_auth_randomizer);
-            let nk = *sk_sender.nullifier_key();
-            let ak = sk_sender.spend_auth_key().into();
-            let mut sct = tct::Tree::new();
-            sct.insert(tct::Witness::Keep, note_commitment).unwrap();
-            let anchor = sct.root();
-            let state_commitment_proof = sct.witness(note_commitment).unwrap();
-            let balance_commitment = value_to_send.commit(v_blinding);
-            let rk: VerificationKey<SpendAuth> = rsk.into();
-            let nf = Nullifier::derive(&nk, 0.into(), &note_commitment);
-
-            let incorrect_nf = Nullifier::derive(&nk, 5.into(), &note_commitment);
-
-            let blinding_r = Fq::rand(&mut OsRng);
-            let blinding_s = Fq::rand(&mut OsRng);
-            let proof = SpendProof::prove(
-                blinding_r,
-                blinding_s,
-                &pk,
-                state_commitment_proof,
-                note,
-                v_blinding,
-                spend_auth_randomizer,
-                ak,
-                nk,
-                anchor,
-                balance_commitment,
-                nf,
-                rk,
-            )
-            .expect("can create proof");
-
-            let proof_result = proof.verify(&vk, anchor, balance_commitment, incorrect_nf, rk);
-            assert!(proof_result.is_err());
-        }
-    }
-
-    proptest! {
-        #![proptest_config(ProptestConfig::with_cases(2))]
-    #[test]
-    /// Check that the `SpendProof` verification fails when using balance
-    /// commitments with different blinding factors.
-    fn spend_proof_verification_balance_commitment_integrity_failure(seed_phrase_randomness in any::<[u8; 32]>(), spend_auth_randomizer in fr_strategy(), value_amount in 2..200u64, v_blinding in fr_strategy(), incorrect_blinding_factor in fr_strategy()) {
-        let (pk, vk) = SpendCircuit::generate_prepared_test_parameters();
-        let mut rng = OsRng;
-
-        let seed_phrase = SeedPhrase::from_randomness(seed_phrase_randomness);
-        let sk_sender = SpendKey::from_seed_phrase(seed_phrase, 0);
-        let fvk_sender = sk_sender.full_viewing_key();
-        let ivk_sender = fvk_sender.incoming();
-        let (sender, _dtk_d) = ivk_sender.payment_address(0u32.into());
-
-        let value_to_send = Value {
-            amount: value_amount.into(),
-            asset_id: asset::Cache::with_known_assets().get_unit("upenumbra").unwrap().id(),
-        };
-
-        let note = Note::generate(&mut rng, &sender, value_to_send);
-        let note_commitment = note.commit();
-        let rsk = sk_sender.spend_auth_key().randomize(&spend_auth_randomizer);
-        let nk = *sk_sender.nullifier_key();
-        let ak = sk_sender.spend_auth_key().into();
-        let mut sct = tct::Tree::new();
-        sct.insert(tct::Witness::Keep, note_commitment).unwrap();
-        let anchor = sct.root();
-        let state_commitment_proof = sct.witness(note_commitment).unwrap();
-        let balance_commitment = value_to_send.commit(v_blinding);
-        let rk: VerificationKey<SpendAuth> = rsk.into();
-        let nf = Nullifier::derive(&nk, 0.into(), &note_commitment);
-
-        let blinding_r = Fq::rand(&mut OsRng);
-        let blinding_s = Fq::rand(&mut OsRng);
-        let proof = SpendProof::prove(
-            blinding_r,
-            blinding_s,
-            &pk,
-            state_commitment_proof,
-            note,
-            v_blinding,
-            spend_auth_randomizer,
-            ak,
-            nk,
-            anchor,
-            balance_commitment,
-            nf,
-            rk,
-        )
-        .expect("can create proof");
-
-        let incorrect_balance_commitment = value_to_send.commit(incorrect_blinding_factor);
-
-        let proof_result = proof.verify(&vk, anchor, incorrect_balance_commitment, nf, rk);
-        assert!(proof_result.is_err());
-    }
-    }
-
-    proptest! {
-            #![proptest_config(ProptestConfig::with_cases(2))]
-        #[test]
-        /// Check that the `SpendProof` verification fails when the incorrect randomizable verification key is used.
-        fn spend_proof_verification_fails_rk_integrity(seed_phrase_randomness in any::<[u8; 32]>(), spend_auth_randomizer in fr_strategy(), value_amount in 2..200u64, v_blinding in fr_strategy(), incorrect_spend_auth_randomizer in fr_strategy()) {
-            let (pk, vk) = SpendCircuit::generate_prepared_test_parameters();
-            let mut rng = OsRng;
-
-            let seed_phrase = SeedPhrase::from_randomness(seed_phrase_randomness);
-            let sk_sender = SpendKey::from_seed_phrase(seed_phrase, 0);
-            let fvk_sender = sk_sender.full_viewing_key();
-            let ivk_sender = fvk_sender.incoming();
-            let (sender, _dtk_d) = ivk_sender.payment_address(0u32.into());
-
-            let value_to_send = Value {
-                amount: value_amount.into(),
-                asset_id: asset::Cache::with_known_assets().get_unit("upenumbra").unwrap().id(),
-            };
-
-            let note = Note::generate(&mut rng, &sender, value_to_send);
-            let note_commitment = note.commit();
-            let rsk = sk_sender.spend_auth_key().randomize(&spend_auth_randomizer);
-            let nk = *sk_sender.nullifier_key();
-            let ak = sk_sender.spend_auth_key().into();
-            let mut sct = tct::Tree::new();
-            sct.insert(tct::Witness::Keep, note_commitment).unwrap();
-            let anchor = sct.root();
-            let state_commitment_proof = sct.witness(note_commitment).unwrap();
-            let balance_commitment = value_to_send.commit(v_blinding);
-            let rk: VerificationKey<SpendAuth> = rsk.into();
-            let nf = Nullifier::derive(&nk, 0.into(), &note_commitment);
-
-            let incorrect_rsk = sk_sender
-                .spend_auth_key()
-                .randomize(&incorrect_spend_auth_randomizer);
-            let incorrect_rk: VerificationKey<SpendAuth> = incorrect_rsk.into();
-
-            let blinding_r = Fq::rand(&mut OsRng);
-            let blinding_s = Fq::rand(&mut OsRng);
-            let proof = SpendProof::prove(
-                blinding_r,
-                blinding_s,
-                &pk,
-                state_commitment_proof,
-                note,
-                v_blinding,
-                spend_auth_randomizer,
-                ak,
-                nk,
-                anchor,
-                balance_commitment,
-                nf,
-                rk,
-            )
-            .expect("should be able to form proof");
-
-            let proof_result = proof.verify(&vk, anchor, balance_commitment, nf, incorrect_rk);
-            assert!(proof_result.is_err());
-        }
-    }
-
-    proptest! {
-            #![proptest_config(ProptestConfig::with_cases(2))]
-        #[test]
-        /// Check that the `SpendProof` verification always suceeds for dummy (zero value) spends.
-        fn spend_proof_dummy_verification_suceeds(seed_phrase_randomness in any::<[u8; 32]>(), spend_auth_randomizer in fr_strategy(), v_blinding in fr_strategy()) {
-            let (pk, vk) = SpendCircuit::generate_prepared_test_parameters();
-            let mut rng = OsRng;
-
-            let seed_phrase = SeedPhrase::from_randomness(seed_phrase_randomness);
-            let sk_sender = SpendKey::from_seed_phrase(seed_phrase, 0);
-            let fvk_sender = sk_sender.full_viewing_key();
-            let ivk_sender = fvk_sender.incoming();
-            let (sender, _dtk_d) = ivk_sender.payment_address(0u32.into());
-
-            let value_to_send = Value {
-                amount: 0u64.into(),
-                asset_id: asset::Cache::with_known_assets().get_unit("upenumbra").unwrap().id(),
-            };
-
-            let note = Note::generate(&mut rng, &sender, value_to_send);
-            let note_commitment = note.commit();
-            let rsk = sk_sender.spend_auth_key().randomize(&spend_auth_randomizer);
-            let nk = *sk_sender.nullifier_key();
-            let ak = sk_sender.spend_auth_key().into();
-            let sct = tct::Tree::new();
-            let anchor = sct.root();
-            let state_commitment_proof = tct::Proof::dummy(&mut OsRng, note_commitment);
-            // Using a random blinding factor here, but the proof will verify
-            // since for dummies we only check if the value is zero, and choose
-            // not to enforce the other equality constraint.
-            let balance_commitment = value_to_send.commit(v_blinding);
-            let rk: VerificationKey<SpendAuth> = rsk.into();
-            let nf = Nullifier::derive(&nk, 0.into(), &note_commitment);
-
-            let blinding_r = Fq::rand(&mut OsRng);
-            let blinding_s = Fq::rand(&mut OsRng);
-            let proof = SpendProof::prove(
-                blinding_r,
-                blinding_s,
-                &pk,
-                state_commitment_proof,
-                note,
-                v_blinding,
-                spend_auth_randomizer,
-                ak,
-                nk,
-                anchor,
-                balance_commitment,
-                nf,
-                rk,
-            )
-            .expect("should be able to form proof");
-
-            let proof_result = proof.verify(&vk, anchor, balance_commitment, nf, rk);
-            assert!(proof_result.is_ok());
-        }
     }
 
     struct MerkleProofCircuit {
