+++ crates/crypto/proof-params/benches/nullifier_derivation.rs
@@ -6,11 +6,11 @@ use ark_relations::r1cs::{
 use decaf377::Fq;
 use penumbra_asset::Value;
 use penumbra_crypto::{
-    proofs::groth16::{NullifierDerivationCircuit, NullifierDerivationProof},
     Note, Nullifier, Rseed,
 };
 use penumbra_keys::keys::{NullifierKey, SeedPhrase, SpendKey};
 use penumbra_proof_params::NULLIFIER_DERIVATION_PROOF_PROVING_KEY;
+use penumbra_shielded_pool::{NullifierDerivationCircuit, NullifierDerivationProof};
 use penumbra_tct as tct;
 
 use criterion::{criterion_group, criterion_main, Criterion};
