+++ tools/parameter-setup/Cargo.toml
@@ -7,6 +7,8 @@ publish = false
 [dependencies]
 penumbra-crypto = { path = "../../crates/core/crypto/" }
 penumbra-dex = { path = "../../crates/core/component/dex/" }
+penumbra-governance = { path = "../../crates/core/component/governance/" }
+penumbra-shielded-pool = { path = "../../crates/core/component/shielded-pool/" }
 penumbra-stake = { path = "../../crates/core/component/stake/" }
 ark-groth16 = "0.4"
 ark-serialize = "0.4"
